Go vs Rust: Article fixes
People on Reddit and Hacker News pointed out a few issues
that I'm happy to fix. I went ahead and corrected/extended
a few sections based on the feedback so far.

I also added two screenshots which I missed to submit yesterday.
Note that I updated the date to today, to reflect that
date in the second screenshot and to avoid confusion.

One function definition was missing, which I added as well.